# docs/api/openapi.core.v1.yaml
openapi: 3.1.0
info:
  title: DELMOS Core First Slice API
  version: 0.1.0-draft
  description: >-
    Target contract for the first Core role and approval slice. Not implemented.
    Session establishment and external integrations are outside this fragment.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /projects/{project_id}/trace-links:
    post:
      operationId: createTraceLink
      summary: Create a typed traceability link between project work products
      x-required-permission: trace.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTraceLink'
      responses:
        '201':
          description: Traceability link created and audited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceLink'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
    get:
      operationId: listTraceLinks
      summary: List project traceability links (Change Impact review list, GRP-03)
      x-required-permission: wp.read
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: suspect
          in: query
          required: false
          description: When "true", return only links flagged is_suspect after a source revision.
          schema: { type: string, enum: ['true', 'false'] }
      responses:
        '200':
          description: Traceability links, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TraceLink' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
  /projects/{project_id}/trace-links/{trace_link_id}/acknowledge:
    post:
      operationId: acknowledgeTraceLink
      summary: Clear the is_suspect flag after reviewer confirmation (GRP-03)
      x-required-permission: trace.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: trace_link_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Suspect flag cleared or already clear (idempotent) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
...
  /projects/{project_id}/traceability/{source_id}:
    get:
      operationId: traverseTraceability
      summary: Traverse the project traceability graph from a work product
      x-required-permission: wp.read
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: source_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Bounded recursive traceability path
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TracePathEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: delmos_session
      description: Server-side HttpOnly session; its creation is outside this fragment.
  parameters:
...
    ProjectId:
      in: path
      name: project_id
      required: true
      schema: { type: string, format: uuid }
...
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Missing permission for a visible action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Hidden:
      description: Resource is missing or not visible in this scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Stale row version or idempotency key reused with different payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Invariant:
      description: Validation, SoD, quorum or lifecycle rule failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    CreateTraceLink:
      type: object
      additionalProperties: false
      required: [source_id, target_id, relation_kind]
      properties:
        source_id: { type: string, format: uuid }
        source_revision_id: { type: string, format: uuid }
        target_id: { type: string, format: uuid }
        target_revision_id: { type: string, format: uuid }
        relation_kind: { type: string, enum: [verifies, satisfies, refines, derives_from] }
    TraceLink:
      type: object
      required: [id, source_id, target_id, relation_kind, is_suspect]
      properties:
        id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        source_revision_id: { type: string, format: uuid }
        target_id: { type: string, format: uuid }
        target_revision_id: { type: string, format: uuid }
        relation_kind: { type: string, enum: [verifies, satisfies, refines, derives_from] }
        is_suspect: { type: boolean }
    TracePathEntry:
      type: object
      required: [depth, source_id, source_code, target_id, target_code, target_type, relation_kind, is_suspect, is_cycle]
      properties:
        depth: { type: integer, minimum: 1 }
        source_id: { type: string, format: uuid }
        source_code: { type: string }
        target_id: { type: string, format: uuid }
        target_code: { type: string }
        target_type: { type: string }
        source_revision_id: { type: string, format: uuid }
        target_revision_id: { type: string, format: uuid }
        relation_kind: { type: string, enum: [verifies, satisfies, refines, derives_from] }
        is_suspect: { type: boolean }
        is_cycle: { type: boolean }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }